update view mnemonic button to flex bottom
