+++ lib/ui/blocs/navigator/navigator_bloc.dart
@@ -69,8 +69,6 @@ class NavigatorBloc extends Bloc<NavigatorEvent, void> {
       _handleNavigateToEditAccount();
     } else if (event is NavigateToUserDetails) {
       _handleNavigateToUserDetails(event);
-    } else if (event is NavigateToShowMnemonicVisualiser) {
-      _handleNavigateToShowMnemonicVisualiser();
     }
   }
 
@@ -181,12 +179,6 @@ class NavigatorBloc extends Bloc<NavigatorEvent, void> {
   void _handleNavigateToUserDetails(NavigateToUserDetails event) {
     _navigatorKey.currentState.push(MaterialPageRoute(builder: (context) {
       return UserDetailsScreen(user: event.user, isMyProfile: false);
-    }));
-  }
-
-  void _handleNavigateToShowMnemonicVisualiser() {
-    _navigatorKey.currentState.push(MaterialPageRoute(builder: (context) {
-      return ShowMnemonicVisualiser();
     }));
   }
 }
