+++ lib/ui/screens/security/export.dart
@@ -2,4 +2,3 @@ export 'security_login_biometrics_screen.dart';
 export 'security_login_password_screen.dart';
 export 'security_set_biometrics_screen.dart';
 export 'security_set_password_screen.dart';
-export 'show_mnemonic_visualiser/index.dart';
