+++ lib/ui/widgets/mnemonic/export.dart
@@ -1,2 +1,3 @@
 export 'export_mnemonic_popup.dart';
 export 'mnemonic_visualizer.dart';
+export 'export_mnemonic_button.dart';
