+++ lib/ui/widgets/mnemonic/export_mnemonic_button.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:mooncake/ui/ui.dart';
+
+class ExportMnemonicButton extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    void _openExportPopup(BuildContext context) {
+      BlocProvider.of<MnemonicBloc>(context).add(ShowExportPopup());
+    }
+
+    return Expanded(
+      child: Align(
+        alignment: Alignment.bottomCenter,
+        child: Container(
+          margin: EdgeInsets.only(bottom: 15),
+          child: PrimaryButton(
+            child: Text(PostsLocalizations.of(context).exportMnemonic),
+            onPressed: () => _openExportPopup(context),
+          ),
+        ),
+      ),
+    );
+  }
+}
